refactor(ui): more SettingsDrawer extraction (network, footer, subnets, vuln, debounced save)
Take SettingsDrawer.tsx from 1771 down to 1149 lines via four new
files plus an in-place auto-save consolidation.

- SettingsDrawerNetworkSection.tsx (~308): IP mode toggle, static IP
  fields, apply button + status message, display-options toggle, and
  VLAN/MTU controls extracted from the Network CollapsibleSection
- SettingsDrawerFooter.tsx (~110): Logs (debug) viewer, Export
  download link, About blurb at the bottom of the drawer
- hooks/useSubnetSettings.ts (~165): subnet list state, form fields,
  fetch/add/toggle/delete callbacks plus the open-time fetch effect
- hooks/useVulnerabilitySettings.ts (~85): vuln scanner settings
  state plus fetch / save callbacks
- hooks/useDebouncedAutoSave.ts (~30): shared debounced auto-save
  hook that replaces eight near-identical useEffect blocks in the
  drawer (thresholds, tests, wifi, link, cable, network discovery,
  snmp, vuln)

SettingsDrawer.tsx still exceeds the 800-line red-flag budget at
1149 lines. The next steps would extract the remaining fetch+save
callbacks (thresholds, IP, tests, iperf suggestions, wifi, network
discovery, snmp, link, cable, log preview) into focused hooks.

Pure relocation, no behavior change.
