+++ ui/src/hooks/useVulnerabilitySettings.ts
@@ -1,0 +1,94 @@
+/**
+ * useVulnerabilitySettings
+ *
+ * Loads and saves CVE/vulnerability scanner settings from
+ * /api/v1/shell/vulnerabilities/settings. Owns the settings shape,
+ * the save-status, and the init/timer refs used by the drawer's
+ * debounced auto-save pattern.
+ */
+
+import { useCallback, useRef, useState } from 'react';
+import { INLINE_DEFAULT_VULNERABILITY_SETTINGS } from '../components/settings/settingsDrawerNormalizer';
+import { LogComponents, logger } from '../lib/logger';
+import type { SaveStatus, VulnerabilityScanSettings } from '../types/settings';
+
+const API_BASE: string = import.meta.env.VITE_API_BASE || '';
+
+interface UseVulnerabilitySettingsResult {
+  vulnSettings: VulnerabilityScanSettings;
+  setVulnSettings: (value: VulnerabilityScanSettings) => void;
+  vulnStatus: SaveStatus;
+  vulnInitRef: React.MutableRefObject<boolean>;
+  vulnTimerRef: React.MutableRefObject<ReturnType<typeof setTimeout> | null>;
+  fetchVulnSettings: () => Promise<void>;
+  saveVulnSettings: () => Promise<void>;
+}
+
+export function useVulnerabilitySettings(): UseVulnerabilitySettingsResult {
+  const [vulnSettings, setVulnSettings] = useState<VulnerabilityScanSettings>(
+    INLINE_DEFAULT_VULNERABILITY_SETTINGS,
+  );
+  const [vulnStatus, setVulnStatus] = useState<SaveStatus>('idle');
+  const vulnInitRef = useRef(true);
+  const vulnTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+
+  const fetchVulnSettings = useCallback(async () => {
+    try {
+      const response = await fetch(`${API_BASE}/api/v1/shell/vulnerabilities/settings`, {
+        credentials: 'include',
+      });
+      if (response.ok) {
+        const data = await (response.json() as Promise<Record<string, unknown>>);
+        setVulnSettings({
+          enabled: data.enabled ?? false,
+          cveDatabase: data.cve_database ?? data.cveDatabase ?? 'nvd',
+          nvdApiKey: data.nvd_api_key ?? data.nvdApiKey ?? '',
+          updateInterval: data.update_interval ?? data.updateInterval ?? 86400,
+          severityThreshold: data.severity_threshold ?? data.severityThreshold ?? 'medium',
+          maxConcurrent: data.max_concurrent ?? data.maxConcurrent ?? 5,
+          autoScan: data.auto_scan ?? data.autoScan ?? false,
+        });
+      }
+    } catch (err) {
+      logger.error(LogComponents.CONFIG, 'Failed to fetch vulnerability settings', err);
+    }
+  }, []);
+
+  const saveVulnSettings = useCallback(async () => {
+    setVulnStatus('saving');
+    try {
+      const response = await fetch(`${API_BASE}/api/v1/shell/vulnerabilities/settings`, {
+        method: 'PUT',
+        headers: { 'Content-Type': 'application/json' },
+        credentials: 'include',
+        body: JSON.stringify({
+          enabled: vulnSettings.enabled,
+          cveDatabase: vulnSettings.cveDatabase,
+          nvdApiKey: vulnSettings.nvdApiKey,
+          updateInterval: vulnSettings.updateInterval,
+          severityThreshold: vulnSettings.severityThreshold,
+          maxConcurrent: vulnSettings.maxConcurrent,
+          autoScan: vulnSettings.autoScan,
+        }),
+      });
+      if (response.ok) {
+        setVulnStatus('saved');
+        setTimeout(() => setVulnStatus('idle'), 2000);
+      } else {
+        setVulnStatus('error');
+      }
+    } catch {
+      setVulnStatus('error');
+    }
+  }, [vulnSettings]);
+
+  return {
+    vulnSettings,
+    setVulnSettings,
+    vulnStatus,
+    vulnInitRef,
+    vulnTimerRef,
+    fetchVulnSettings,
+    saveVulnSettings,
+  };
+}
